backend/requirements.txt
fastapi>=0.115.12
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
psycopg2-binary>=2.9.9
alembic>=1.15.2
email-validator>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.1
argon2-cffi>=23.1.0
